backend/requirements.txt
fastapi
uvicorn
numpy
httpx
pytest
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
